upstream

+++ source/game_sa/Events/EventVehicleHitAndRun.cpp
@@ -38,8 +38,8 @@ void CEventVehicleHitAndRun::ReportCriminalEvent(CPed* ped)
 {
     if (IsCriminalEvent()) {
         if (m_victim->m_nPedType == PED_TYPE_COP)
-            FindPlayerWanted()->RegisterCrime(eCrimeType::CRIME_KILL_COP_PED_WITH_CAR, m_vehicle->GetPosition(), m_vehicle->m_pDriver, false);
+            FindPlayerWanted()->RegisterCrime(eCrimeType::CRIME_KILL_COP_PED_WITH_CAR, m_vehicle->GetPosition(), (uint32)m_vehicle->m_pDriver, false);
         else
-            FindPlayerWanted()->RegisterCrime(eCrimeType::CRIME_KILL_PED_WITH_CAR, m_vehicle->GetPosition(), m_vehicle->m_pDriver, false);
+            FindPlayerWanted()->RegisterCrime(eCrimeType::CRIME_KILL_PED_WITH_CAR, m_vehicle->GetPosition(), (uint32)m_vehicle->m_pDriver, false);
     }
 }
